move build job to build worker and kvm test to kvmtest workers (#113)
there will be three roles of workers
- jenkins-build-workers
- jenkins-vstest-workers
- jenkins-kvmtest-workers

The build workers are mainly for build the image.
The vstest workers are mainly for swss/sonic-utilities vs tests
The kvmtest workers are manly for sonic-mgmt kvm tests

The vstest and kvmtest can share the same physical servers. However, to help debug, we introduce two roles so that they can be separated to avoid their jobs affecting each other.