[utilities] Remove dependencies on latest swss and sairedis updates from utilities builds (#146)
Currently the builds for sonic-utilities fetch and install the very latest versions of swss-common, sairedis, and swss to run the tests.

This is problematic because if ever any of those packages are out of sync (e.g. swss-common receives a new update and the other two have not been rebuilt in the interim to accommodate this new change), then the utilities tests will experience a bunch of odd errors/failures in a very short window of time. These are difficult to debug and they block the PR pipeline for hours at a time.

We could get around this by rebuilding all of the packages for the utilities build, but seeing as the utilities submodule can be updated and merged into the master build-image independently of the other three I think this is unnecessary.

We should install the newly built utilities package into docker-sonic-vs and verify that that works. This verifies that a PR/given build of utilities is compatible with the virtual switch image and that the submodule can be updated without any issues.

Signed-off-by: Danny Allen <daall@microsoft.com>