[sonic-utilities] Test and build Python 3 package, but don't fail build yet (#185)
In preparation for moving sonic-utilities to Python 3, in the sonic-utilities build script, we install all the Python 3 dependencies and attempt to test and build the Python 3 version of the package. This will ensure we can see the Python 3 build/test success status in the PR check build output. However, we will not fail the build job yet if the Python 3 test or build fails.